requirements.txt
ultralytics
fastapi
uvicorn
opencv-python
pillow
webcolors
scipy
python-multipart
